Slugify complété et déporté dans son propre service
